requirements.txt
atlassian-python-api==3.15.0
python-dotenv==0.19.2
